Transformed some parts of the code into different components
