+++ src/components/loading.js
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const LoadingAnimation = () => (
+  <div className="loadingDiv">
+    <h2 data-testid="loading" className="loading">Loading, please wait</h2>
+    <div className="loadingio-spinner-reload-zln0owzosys">
+      <div className="ldio-cf5nk0nv464">
+        <div>
+          <div />
+          <div />
+          <div />
+        </div>
+      </div>
+    </div>
+  </div>
+);
+
+export default LoadingAnimation;
