+++ src/index.css
@@ -18,6 +18,19 @@ body {
   background-color: #5788e4;
 }
 
+.loadingDiv {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-top: 25%;
+}
+
+.loading {
+  align-items: center;
+  color: #dde4f6;
+}
+
 @keyframes ldio-cf5nk0nv464 {
   0% {
     transform: rotate(0);
